# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "SmartDNS Control Plane API",
  "version": "1.0.0",
  "description": "Публичный API панели управления. Веб-интерфейс использует ровно эти же эндпоинты — приватного канала нет.\n\n**Аутентификация.** Сессионная cookie `sdns_session` (HttpOnly, SameSite=Lax) плюс заголовок `X-CSRF-Token` на всех изменяющих методах. Токен возвращают `POST /auth/login` и `GET /auth/me`.\n\n**Одновременные изменения.** Изменяемые объекты содержат поле `version`. Передайте его в `If-Match`, чтобы получить 412 вместо перезаписи чужих изменений.\n\n**Повторные попытки.** `Idempotency-Key` на регистрации нод, сборке, выкате, откате и ручном обновлении списков: повтор возвращает сохранённый ответ, а не выполняет действие второй раз.\n\n**Время** — RFC 3339 в UTC. **Постраничность** — курсорная.\n\nAPI агентов (`/agent/v1/…`) обслуживается отдельным слушателем со взаимным TLS и здесь не описан: он предназначен только для нод. Его контракт — в `docs/API.md`.",
  "license": {
   "name": "См. LICENSE"
  }
 },
 "servers": [
  {
   "url": "/api/v1",
   "description": "Тот же origin, что и веб-панель"
  }
 ],
 "security": [
  {
   "sessionCookie": []
  }
 ],
...
 "paths": {
...
  "/rule-sets": {
   "get": {
    "tags": [
     "Наборы правил"
    ],
    "summary": "Список наборов",
    "description": "Включает имена встроенных пресетов.",
    "responses": {
     "200": {
      "description": "Список",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "properties": {
          "items": {
           "type": "array",
           "items": {
            "$ref": "#/components/schemas/RuleSet"
           }
          }
         },
         "required": [
          "items"
         ]
        }
       }
      }
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "Наборы правил"
    ],
    "summary": "Создать набор",
    "description": "",
    "responses": {
     "201": {
      "description": "Создано",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RuleSet"
        }
       }
      }
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleSetRequest"
       }
      }
     }
    }
   }
  },
  "/rule-sets/{id}": {
   "get": {
    "tags": [
     "Наборы правил"
    ],
    "summary": "Набор",
    "description": "Источники, история версий, журнал загрузок и предпросмотр активных правил.",
    "responses": {
     "200": {
      "description": "Успех"
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ]
   },
   "patch": {
    "tags": [
     "Наборы правил"
    ],
    "summary": "Изменить набор",
    "description": "",
    "responses": {
     "200": {
      "description": "Обновлено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/RuleSet"
        }
       }
      }
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     },
     {
      "name": "If-Match",
      "in": "header",
      "schema": {
       "type": "string"
      },
      "description": "Версия объекта из поля version. Защищает от перезаписи чужих изменений."
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RuleSetRequest"
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "Наборы правил"
    ],
    "summary": "Удалить набор",
    "description": "Отклоняется с 409, если набор используется сервисом.",
    "responses": {
     "200": {
      "description": "Успех"
     },
     "400": {
      "description": "Некорректный запрос",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "401": {
      "description": "Не аутентифицирован",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "403": {
      "description": "Недостаточно прав или не пройдена CSRF-проверка",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "404": {
      "description": "Не найдено",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "409": {
      "description": "Конфликт: объект используется или изменён другим клиентом",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "412": {
      "description": "Версия объекта устарела (If-Match)",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     },
     "422": {
      "description": "Конфигурация не компилируется",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Error"
        }
       }
      }
     }
    },
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "format": "uuid"
      }
     }
    ]
   }
  },
...
 },
 "components": {
  "securitySchemes": {
   "sessionCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "sdns_session"
   }
  },
  "schemas": {
   "Error": {
    "type": "object",
    "required": [
     "code",
     "message",
     "request_id"
    ],
    "properties": {
     "code": {
      "type": "string",
      "examples": [
       "invalid_request"
      ]
     },
     "message": {
      "type": "string"
     },
     "details": {
      "description": "Структурированный контекст: список зависимостей при 409 или список конфликтов правил."
     },
     "request_id": {
      "type": "string"
     }
    }
   },
...
   "RuleSet": {
    "type": "object",
    "properties": {
     "id": {
      "type": "string",
      "format": "uuid"
     },
     "name": {
      "type": "string"
     },
     "update_mode": {
      "type": "string",
      "enum": [
       "auto_apply",
       "manual_approve",
       "manual_only"
      ]
     },
     "interval_sec": {
      "type": "integer"
     },
     "allow_regex": {
      "type": "boolean"
     },
     "entry_count": {
      "type": "integer"
     },
     "active_sequence": {
      "type": [
       "integer",
       "null"
      ]
     },
     "active_hash": {
      "type": [
       "string",
       "null"
      ]
     },
     "pending_version_id": {
      "type": [
       "string",
       "null"
      ]
     },
     "version": {
      "type": "integer"
     }
    }
   },
   "RuleSetRequest": {
    "type": "object",
    "required": [
     "name"
    ],
    "properties": {
     "name": {
      "type": "string"
     },
     "description": {
      "type": "string"
     },
     "update_mode": {
      "type": "string",
      "enum": [
       "auto_apply",
       "manual_approve",
       "manual_only"
      ]
     },
     "interval_sec": {
      "type": "integer",
      "minimum": 300
     },
     "allow_regex": {
      "type": "boolean",
      "default": false,
      "description": "RE2, линейное время, длина не больше 256 символов."
     },
     "priority": {
      "type": "integer"
     },
     "manual_include": {
      "type": "array",
      "items": {
       "type": "string"
      }
     },
     "manual_exclude": {
      "type": "array",
      "items": {
       "type": "string"
      }
     }
    }
   },
...
  }
 }
}